requirements.txt
google-generativeai>=0.8.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0

faiss-cpu>=1.8.0          # For offline vector search (or swap with chromadb if you prefer)
numpy>=1.26
tqdm>=4.66

duckduckgo-search>=6.2.0  # Example online search provider (free, no key)
# OR Tavily:
# tavily-python>=0.3.1

python-dotenv>=1.0.1      # Optional: load .env

rich>=13.9.0              # Nice CLI output (optional)
duckduckgo_search
ddgs
tavily